Style: 컴포넌트 내부 스타일코드 분리
