+++ src/components/Categories.tsx
@@ -1,4 +1,4 @@
-import styled from "styled-components";
+import * as S from "./CategoriesStyles";
 
 interface CategoriesProps {
   categories: string[];
@@ -6,39 +6,18 @@ interface CategoriesProps {
   onClick: (category: string) => void;
 }
 
-interface CategoryButtonProps {
-  readonly $isActive: boolean;
-}
-
 export default function Categories({ categories, selected, onClick }: CategoriesProps) {
   return (
-    <Container>
+    <S.Container>
       {categories.map((category) => (
-        <CategoryButton
+        <S.CategoryButton
           key={category}
           onClick={() => onClick(category)}
           $isActive={selected === category}
         >
           {category}
-        </CategoryButton>
+        </S.CategoryButton>
       ))}
-    </Container>
+    </S.Container>
   );
 }
-
-const Container = styled.ul`
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-`;
-
-const CategoryButton = styled.button<CategoryButtonProps>`
-  padding: 0.5rem 0.75rem;
-  border-radius: 5px;
-  border: 1px solid var(--color-primary);
-  background: ${({ $isActive }) => ($isActive ? "var(--color-primary)" : "var(--color-white)")};
-  color: ${({ $isActive }) => ($isActive ? "var(--color-white)" : "black")};
-  font-size: 1rem;
-  line-height: 1.2rem;
-  cursor: pointer;
-`;
