+++ src/components/OptionsStyles.ts
@@ -1,0 +1,21 @@
+import styled from "styled-components";
+
+export const Container = styled.ul`
+  display: flex;
+  gap: 12px;
+`;
+
+export const OptionsContainer = styled.div`
+  display: flex;
+  align-items: flex-start;
+  gap: 4px;
+  cursor: pointer;
+`;
+
+export const Option = styled.button`
+  padding: 0;
+  background: none;
+  border: none;
+  font-size: 0.875rem;
+  color: #9fa6b2;
+`;
