+++ src/components/button/ModalButton.tsx
@@ -1,18 +1,4 @@
-import styled from "styled-components";
-import { MainButton } from "./Basic";
-import { breakPoints } from "@/common/media";
-
-const ModalBasicButton = styled(MainButton)<{ $isDelete: boolean }>`
-  padding: 1rem 1.25rem;
-  border-radius: 0.5rem;
-  font-size: 1rem;
-
-  background: ${({ $isDelete }) => ($isDelete ? "var(--color-red)" : "")};
-
-  @media only screen and (${breakPoints.mobile}) {
-    width: 100%;
-  }
-`;
+import { ModalBasicButton } from "./ButtonStyles";
 
 interface ModalButtonProps {
   action: string;
