+++ src/components/card/Card.module.css
@@ -89,7 +89,6 @@
   width: 100%;
   display: flex;
   flex-direction: column;
-  /* gap: 10px; */
 }
 
 .info {
