+++ src/components/inputs/InputsStyles.ts
@@ -1,5 +1,17 @@
 import styled from "styled-components";
 
+/** hero 링크추가 input */
+export const AddLinkInput = styled.input`
+  padding: 1rem 2.5rem;
+  width: 100%;
+  background: var(--color-white);
+  border: 1px solid var(--color-primary);
+  border-radius: 1rem;
+  line-height: 1.5rem;
+  outline-style: none;
+`;
+
+/** 회원가입, 로그인 페이지 input */
 export const Container = styled.div`
   width: 350px;
   display: flex;
