+++ src/components/inputs/SearchbarStyles.ts
