+++ src/components/modal/DeleteFolder.tsx
@@ -1,4 +1,5 @@
-import styled from "styled-components";
+import * as S from "./modalStyles/DeleteFolderStyles";
+
 import ModalTitle from "./ModalTitle";
 import ModalButton from "@/components/button/ModalButton";
 
@@ -9,32 +10,12 @@ interface DeleteFolderProps {
 
 export default function DeleteFolder({ currentFolderName, label }: DeleteFolderProps) {
   return (
-    <Contents>
-      <Description>
+    <S.Contents>
+      <S.Description>
         <ModalTitle label={`${label} 삭제`} />
-        <Info>{currentFolderName}</Info>
-      </Description>
+        <S.Info>{currentFolderName}</S.Info>
+      </S.Description>
       <ModalButton action="delete" label="삭제하기" />
-    </Contents>
+    </S.Contents>
   );
 }
-
-const Contents = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 1.5rem;
-`;
-
-const Description = styled.div`
-  width: 100%;
-  text-align: center;
-`;
-
-const Info = styled.p`
-  margin: 0;
-  padding-top: 0.5rem;
-  font-size: 0.875rem;
-  color: var(--color-gray);
-`;
