+++ src/components/modal/EditFolder.tsx
@@ -1,4 +1,5 @@
-import styled from "styled-components";
+import * as S from "./modalStyles/EditFolderStyles";
+
 import ModalTitle from "./ModalTitle";
 import ModalButton from "@/components/button/ModalButton";
 
@@ -10,31 +11,10 @@ export default function EditFolder({ currentFolderName }: EditFolderProps) {
   return (
     <>
       <ModalTitle label="폴더 이름 변경" />
-      <Contents>
-        <Input placeholder="내용 입력" defaultValue={currentFolderName} />
+      <S.Contents>
+        <S.Input placeholder="내용 입력" defaultValue={currentFolderName} />
         <ModalButton action="change" label="변경하기" />
-      </Contents>
+      </S.Contents>
     </>
   );
 }
-
-const Contents = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 1rem;
-`;
-
-const Input = styled.input`
-  width: 100%;
-  padding: 1.125rem 1rem;
-  border-radius: 0.5rem;
-  border: 1px solid var(--color-gray-20);
-  outline-style: none;
-  font-size: 1rem;
-
-  &:focus {
-    border: 1px solid var(--color-primary);
-  }
-`;
