+++ src/components/modal/modalStyles/DeleteFolderStyles.ts
@@ -1,0 +1,21 @@
+import styled from "styled-components";
+
+export const Contents = styled.div`
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 1.5rem;
+`;
+
+export const Description = styled.div`
+  width: 100%;
+  text-align: center;
+`;
+
+export const Info = styled.p`
+  margin: 0;
+  padding-top: 0.5rem;
+  font-size: 0.875rem;
+  color: var(--color-gray);
+`;
