+++ src/components/modal/modalStyles/ModalContainerStyles.ts
@@ -1,0 +1,37 @@
+import styled from "styled-components";
+
+export const Container = styled.section`
+  width: 100%;
+  height: 100%;
+  position: fixed;
+  top: 0;
+  left: 0;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  background-color: rgba(0, 0, 0, 0.4);
+  z-index: 1;
+`;
+
+export const Modal = styled.div`
+  position: relative;
+  width: 22.5rem;
+  padding: 2rem 2.5rem;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 1.5rem;
+  background: var(--color-white);
+  border-radius: 1rem;
+  border: 1px solid var(--color-gray-20);
+`;
+
+export const Icon = styled.img`
+  position: absolute;
+  top: 0;
+  right: 0;
+  padding: 1rem;
+  cursor: pointer;
+`;
